TRADE BEACON UPDATE
Adds trade beacons to persistence, devices scattered across the overmap that give access to unique import/export opportunities. This includes a new network device that lets networks take control of trade beacons. The entire system is in place and functional, but we still need to add imports/exports and a variety of trade beacons. I will save that for future updates.