Save error tolerance admin verb
Added admin verb to the server tab to allow changing at runtime the error tolerance for saving the world. Allow skipping over recoverable and unrecoverable error if wanted. Meant to be a sort of last-resort tool to force a save if something on the map breaks the save in a limited way.