Save logging + Fixes (#473)
* Add save logging + cutup save code

* Save error tolerance admin verb

Added admin verb to the server tab to allow changing at runtime the error tolerance for saving the world. Allow skipping over recoverable and unrecoverable error if wanted. Meant to be a sort of last-resort tool to force a save if something on the map breaks the save in a limited way.

* Update misc.dm

* Reorganized and moved saving code

Renamed and moved the code for saving in SSpersistence over to it's own file since it was pretty massive and hard to navigate.

* Update persistence.dm

* Moved save statistics to own file

* Moved limbo stuff to it's own file

* Also renamed "DeserializeOneOff" to "LoadFromLimbo" to match the other limbo proc names.

* Began moving save loading code to new file

* Update persistence.dm

* Fix mimic_edge saving

* fixed save text being hard to read in darkmode

* Moved and cut up world loading code

* Renamed and duplicated exception filters, one for saving, and one for loading. Since loading doesn't need to unblock entering or restart subsystems on an exception.
* Cut up loading code after applying some of the fixes upstream.

* TEMPORARY fix for debugger not working

* Formatting

* Fix map not loading at the right zlevel

* Make planetoid_data load properly

* Update check-paths.sh

* Revert "TEMPORARY fix for debugger not working"

This reverts commit cee6ea4a90d44022c6faf20ea686eee9cad8dac2.

* Move last save display to persistent modpack

* Fix loading save not filling saved_levels

level_data.setup_level_data() is not called after loading level data from a save. But that's where it was adding itself to the list of saved levels. (which really should have stayed inside ssmapping for this very reason)

* Removed duplicate cleanup

* Throw the save/load exception if any

Silently catching it means we don't get a call stack and nice things like that. So, I'm rethrowing it for the runtime log to pick it up.

* Cleanup save debug objects a bit

* changed comments

* Added configuration hooks for persistence configs

* Added the persistence config options to the example config file.
* Renamed an ambiguous config options for persistence.

* Used a stub system for config

Removes dependency on the persistence modpack.

* updated config example

* Clarified comment on autosave config